complete ex 4.21

+++ part4/controllers/blogs.js
@@ -30,9 +30,13 @@ blogsRouter.post('/', async (request, response, next) => {
 
   const user = await User.findById(decodedToken.id);
 
+  if (!user) {
+    return response.status(404).end();
+  }
+
   const blog = new Blog(body);
 
-  blog.user = user.id;
+  blog.user = user._id;
 
   if (body.likes === undefined) {
     blog.likes = 0;
@@ -53,8 +57,30 @@ blogsRouter.post('/', async (request, response, next) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' });
+  }
+
+  const user = await User.findById(decodedToken.id);
+
+  if (!user) {
+    return response.status(404).end();
+  }
+
+  const blog = await Blog.findById(request.params.id);
+
+  if (!blog) {
+    return response.status(404).end();
+  }
+
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id);
+    response.status(204).end();
+  } else {
+    return response.status(401).json({ error: 'invalid user' });
+  }
 });
 
 blogsRouter.put('/:id', async (request, response) => {
